fix: not re-writing ojcore due to testing watching

+++ index.js
@@ -14,41 +14,27 @@ module.exports = class ojPlugin {
     }
   }
   apply(compiler) {
-    const folder = compiler.options.context;
-    const ojCore = path.join(folder, this.ojCorePath);
+    const ojCore = path.join(process.cwd(), this.ojCorePath);
     let originalOjCore = fs.readFileSync(ojCore, 'utf8');
-  
-    const newOjCore = originalOjCore
-                                    // avoid un-resolvable dynamic require
-                                    .replace("require(requestedBundles,", "myUniqueFunctionToReplaceInOjCore(requestedBundles,")
-                                    // we don't want 'oj' to be added in window
-                                    .replace("typeof window !== 'undefined'", "false")
-                                    // self equals to window, so we replace it for a new object
-                                    .replace("_scope = self;", "_scope = {};")
-                                    // if amd is present, we define the ojs/ojcore module returning our oj object
-                                    .replace(";return oj;", `
-                                              if(oj.__isAmdLoaderPresent()) {
-                                                define("ojs/ojcore", [], function() {
-                                                  return oj;
-                                                } );
-                                              }
-                                              ;return oj;`);
-
-    fs.writeFileSync(ojCore, newOjCore);
-
-    compiler.plugin("compilation", function (compilation) {
-      //the main compilation instance
-      //all subsequent methods are derived from compilation.plugin
-      compilation.plugin("optimize-chunk-assets", function (chunks, callback) {
-        //ToDo this seems a pretty non-performant approach, need to figure out how to improve it
-        chunks.forEach(function (chunk) {
-          chunk.files.forEach(function (file) {
-            compilation.assets[file] = new ConcatSource(compilation.assets[file].source().replace("myUniqueFunctionToReplaceInOjCore", "require"));
-          });
-        });
-        fs.writeFileSync(ojCore, originalOjCore);
-        callback();
-      });
-    });
+    if(originalOjCore.indexOf('myUniqueFunctionToReplaceInOjCore') === -1) {
+      
+      const newOjCore = originalOjCore
+                                      // avoid un-resolvable dynamic require
+                                      .replace("require(requestedBundles,", "myUniqueFunctionToReplaceInOjCore(requestedBundles,")
+                                      // we don't want 'oj' to be added in window
+                                      .replace("typeof window !== 'undefined'", "false")
+                                      // self equals to window, so we replace it for a new object
+                                      .replace("_scope = self;", "_scope = {};")
+                                      // if amd is present, we define the ojs/ojcore module returning our oj object
+                                      .replace(";return oj;", `
+                                                if(oj.__isAmdLoaderPresent()) {
+                                                  define("ojs/ojcore", [], function() {
+                                                    return oj;
+                                                  } );
+                                                }
+                                                ;return oj;`);
+      
+      fs.writeFileSync(ojCore, newOjCore);
+    }
   }
 };
